+++ src/db/piano/main.js
@@ -1,5 +1,0 @@
-export default {
-  name: 'piano',
-  keys: 88,
-  fretsOnChord: null,
-};