+++ src/db/piano/tunings.js
@@ -1,3 +1,0 @@
-export default {
-  standard: ['C4', 'D4', 'E4', 'F4', 'G4', 'A4', 'B4', 'C5'],
-};